ISPN-10802 Add sizeAsync method

+++ core/src/main/java/org/infinispan/interceptors/impl/CacheLoaderInterceptor.java
@@ -314,7 +314,7 @@ public class CacheLoaderInterceptor<K, V> extends JmxStatsCommandInterceptor imp
          if ((Integer) rv == -1) {
             return super.visitSizeCommand(rCtx, rCommand);
          }
-         return rv;
+         return ((Integer) rv).longValue();
       });
 
    }
